// frontend/src/app/models/rezultatGrupni.ts
export class RezultatGrupni{
    id: number;
    sport: string;
    pol: string;
    format: number;
    ekipe: string[];
    grupaA: number[][];
    grupaB: number[][];

    prosliA: string[];
    prosliB: string[];
    f_4: number[];
    prosliCetvrt: string[];
    f_2: number[];
    prosliPola: string[];
    f: number[];
    mecZaTrece: number[];

    zavrsenoA: number;
    zavrsenoB: number;
    zavrseno4: number;
    zavrseno2: number;
    zavrseno1: number;
}