// frontend/src/app/models/rezultatTenis.ts
export class RezultatTenis{
    id: number;
    pol: string;
    format: number;
    ekipe8: string[];
    ekipe4: string[];
    ekipe3: string[];
    ekipe2: string[];
    ekipe1: string[];
    ekipe123: string[];
    f_8: number[];
    f_4: number[];
    f_3: number[];
    f_2: number[];
    f_1: number[];
    zavrseno8: number;
    zavrseno4: number;
    zavrseno3: number;
    zavrseno2: number;
    zavrseno1: number;

    brTakmicara: number;
}